requirements.txt
aiogram
mistralai
logging
beanie
python-dotenv